+++ boilingcore/config_test.go
@@ -71,6 +71,76 @@ func TestConvertAliases(t *testing.T) {
 	}
 }
 
+func TestConvertAliasesAltSyntax(t *testing.T) {
+	t.Parallel()
+
+	var intf interface{} = map[string]interface{}{
+		"tables": []interface{}{
+			map[string]interface{}{
+				"name":          "table_name",
+				"up_plural":     "a",
+				"up_singular":   "b",
+				"down_plural":   "c",
+				"down_singular": "d",
+
+				"columns": []interface{}{
+					map[string]interface{}{
+						"name":  "a",
+						"alias": "b",
+					},
+				},
+				"relationships": []interface{}{
+					map[string]interface{}{
+						"name":    "ib_fk_1",
+						"local":   "a",
+						"foreign": "b",
+					},
+				},
+			},
+		},
+	}
+
+	aliases := ConvertAliases(intf)
+
+	if len(aliases.Tables) != 1 {
+		t.Fatalf("should have one table alias: %#v", aliases.Tables)
+	}
+
+	table := aliases.Tables["table_name"]
+	if table.UpPlural != "a" {
+		t.Error("value was wrong:", table.UpPlural)
+	}
+	if table.UpSingular != "b" {
+		t.Error("value was wrong:", table.UpSingular)
+	}
+	if table.DownPlural != "c" {
+		t.Error("value was wrong:", table.DownPlural)
+	}
+	if table.DownSingular != "d" {
+		t.Error("value was wrong:", table.DownSingular)
+	}
+
+	if len(table.Columns) != 1 {
+		t.Error("should have one column")
+	}
+
+	if table.Columns["a"] != "b" {
+		t.Error("column alias was wrong")
+	}
+
+	if len(aliases.Tables) != 1 {
+		t.Fatal("should have one relationship alias")
+	}
+
+	rel := table.Relationships["ib_fk_1"]
+	if rel.Local != "a" {
+		t.Error("value was wrong:", rel.Local)
+	}
+	if rel.Foreign != "b" {
+		t.Error("value was wrong:", rel.Foreign)
+	}
+}
+
 func TestConvertTypeReplace(t *testing.T) {
 	t.Parallel()