+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add alternative syntax that's case-sensitive for defining aliases. This is
+  particularly useful if toml key syntax is not good enough or viper keeps
+  lowercasing all your keys (and you have uppercase names in say mssql).
+
 ## [v3.0.0-rc7]
 
 ### Fixed